shift generator to awsconfigfile package
