+++ go.mod
@@ -48,7 +48,7 @@ require (
 	github.com/aws/aws-sdk-go-v2/service/iam v1.18.25
 	github.com/aws/aws-sdk-go-v2/service/internal/presigned-url v1.7.0 // indirect
 	github.com/aws/smithy-go v1.13.5
-	github.com/common-fate/awsconfigfile v0.2.0
+	github.com/common-fate/awsconfigfile v0.4.0
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
 	github.com/danieljoos/wincred v1.1.2 // indirect
 	github.com/dvsekhvalnov/jose2go v1.5.0 // indirect
@@ -65,7 +65,7 @@ require (
 	github.com/olekukonko/tablewriter v0.0.5
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/stretchr/testify v1.8.1
-	golang.org/x/sync v0.1.0
+	golang.org/x/sync v0.1.0 // indirect
 	golang.org/x/sys v0.4.0
 	golang.org/x/term v0.4.0 // indirect
 	golang.org/x/text v0.6.0
